revert DELEGATION_IN_PROGRESS records to DELEGATION_QUEUE (#351)
Closes: #XXX

## Context and purpose of the change

`OnTimeoutPacket` reverts a single record in a pending state; we need to revert multiple pending records in some cases. This PR adds the ability to reset batches of records from pending -> queue state.

Note: the controller module in ibcv3 does not call `OnChanCloseConfirm`, so while this would be the most natural place to put this logic, it is not an option unless we fork or upgrade ibc (to a different version). For now, I put this logic in `RestoreInterchainAccount` (which should also simplify remedying future issues), although we could also put this logic in `OnTimeoutPacket`.


## Brief Changelog

- revert records to queue state when restoring an ICA

## Testing

Configure the go relayer to only relay transfer packets like so
```
paths:
  stride-gaia:
    src:
      chain-id: STRIDE
    dst:
      chain-id: GAIA
    src-channel-filter:
      rule: "allowlist"
      channel-list: [channel-0]
```

Once the chains are up and running, pause hermes, then liquid stake a few times. This will cause records to build up in the DELEGATION_PENDING state.

Then, unpause hermes. This will cause a timeout and channel closure.

Finally, restore the ICA like so
```
$STRIDE_MAIN_CMD tx stakeibc restore-interchain-account GAIA DELEGATION --keyring-backend test --from val1 -y --chain-id $STRIDE_CHAIN_ID
```
This reverts the pending records to the queue state.

## Author's Checklist

I have...

- [ ] Run and PASSED locally all GAIA integration tests
- [ ] If the change is contentful, I either:
    - [ ] Added a new unit test OR 
    - [ ] Added test cases to existing unit tests
- [ ] OR this change is a trivial rework / code cleanup without any test coverage

If skipped any of the tests above, explain.


## Reviewers Checklist

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

I have...

- [ ] reviewed state machine logic
- [ ] reviewed API design and naming
- [ ] manually tested (if applicable)
- [ ] confirmed the author wrote unit tests for new logic
- [ ] reviewed documentation exists and is accurate


## Documentation and Release Note

  - [ ] Does this pull request introduce a new feature or user-facing behavior changes? 
  - [ ] Is a relevant changelog entry added to the `Unreleased` section in `CHANGELOG.md`?
  - [ ] This pull request updates existing proto field values (and require a backend and frontend migration)? 
  - [ ] Does this pull request change existing proto field names (and require a frontend migration)?
  How is the feature or change documented? 
      - [ ] not applicable
      - [ ] jira ticket `XXX` 
      - [ ] specification (`x/<module>/spec/`) 
      - [ ] README.md 
      - [ ] not documented