Fix bug in ICA SendTx (#812)
Closes: #XXX

## Context
### The Problem
With ibc-go v7, the keeper function SendTx was deprecated in favor of the msg_server function...
```go
// from keeper function
k.ICAControllerKeeper.SendTx(...) 

// to msg_server function
msg := icacontrollertypes.NewMsgSendTx(...)
msgServer.SendTx(ctx, msg)
```
Whereas the timestamp in the keeper function was an _absolute_ timestamp (i.e. the timeout in unix nano), the timestamp that should be passed into `NewMsgSendTx` is a _relative_ timestamp (i.e. the offset in nano seconds between the current time and the timeout timestamp).

As the code lives on main, the absolute timestamp is being passed into `NewMsgSendTx`, which is causing the end timestamp to be essentially doubled:
```bash
# Ex: 
Current Time: 1000
Desired Timeout: 1010
NewMsgSendTx(1010) → Returns absolute timestamp of 2010
```

### The Solution
There are two solutions here:
 1. Continue to use the new msg server API and adjust all invocations of SubmitTxs to pass the relative timestamp instead of the absolute timestamp
 2. Revert back to the legacy API
 
Considering option (1) would involve a big refactor, I think it makes sense to move forward with option (2) in the short term.

## Brief Changelog
* Reverted ICA SendTx call to legacy api