reset delegation change in progress to 0 when a channel is restored (#1074)
Closes: #XXX

## Context 
The `DelegationChangesInProgress` field on the validator struct is a mechanism to prevent ICA/ICQ overlap. When ICAs are submitted, the `DelegationChangesInProgress` field is incremented, and when the ack comes back, it is decremented.

In order to prevent ICA/ICQ overlap, when a slash ICQ is relayed, it [[checks](https://github.com/Stride-Labs/stride/blob/main/x/stakeibc/keeper/icqcallbacks_delegator_shares.go#L147-L148)] if there are any active delegation changes in progress, and if so, the query is retried (since the results can’t be trusted)

However, when an ICA packet times out, the channel is closed and the pending packets are never relayed. Since we never receive the ack for these packets, the `DelegationChangesInProgress` field never gets decremented. 

This PR sets the `DelegationChangesInProgress` field to 0 when the ICA channel is restored.

## Brief Changelog
* Set the `DelegationChangesInProgress` to 0 when a channel is restored and added unit test